requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.4
django-countries==7.3.2
django-js-asset==2.0.0
django-mptt==0.13.4
djlint==0.7.6
idna==3.3
importlib-metadata==4.11.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.9.2
tqdm==4.64.0
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
